修复两个问题：1) 超长的名字会被 strip 掉 2) 浮点数能被正常地传输

+++ usmooth/usmooth.app/Pages/Realtime/Realtime.Net.cs
@@ -59,7 +59,7 @@ namespace usmooth.app.Pages
                     m.Name = c.ReadString();
                     m.VertCnt = c.ReadInt32();
                     m.MatCnt = c.ReadInt32();
-                    m.Size = (float)c.ReadInt32();
+                    m.Size = (float)c.ReadFloat();
                     meshes.Add(m);
                 }
             }
